아군기 제외 나머지는 모두 각도를 기반으로 움직이게끔 수정.

+++ src/main/java/game/unit/bullet/Bullet.java
@@ -5,7 +5,7 @@ import game.stage.Stage;
 import game.unit.Unit;
 
 public abstract class Bullet extends Unit {
-    public Bullet(Point pos, int width, int height, int direction) {
+    public Bullet(Point pos, int width, int height, double direction) {
         super(new Point(pos.x - width / 2, pos.y), direction);
     }
